Update and optimise parseEmoji function
